Separate actual camera device start and frame delivery
These two tasks (starting the camera device, and starting working thread that pulls frames
from the started camera device) should be clearly separated, and should not be combined in
one method (as it was with the 'startCapturing' method).

+++ tools/emulator/system/camera/EmulatedQemuCamera.cpp
@@ -84,7 +84,7 @@ status_t EmulatedQemuCamera::Initialize(const char* device_name,
     if (c == NULL) {
         strncpy(first_dim, frame_dims, sizeof(first_dim));
         first_dim[sizeof(first_dim)-1] = '\0';
-    } else if ((c - frame_dims) < sizeof(first_dim)) {
+    } else if (static_cast<size_t>(c - frame_dims) < sizeof(first_dim)) {
         memcpy(first_dim, frame_dims, c - frame_dims);
         first_dim[c - frame_dims] = '\0';
     } else {
